# services/gen3_embeddings/docs/autogenerated_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gen3 Embeddings Service",
    "version": "0.1.2"
  },
  "paths": {
    "/vectorstore/collections/{collection_name}/embeddings/{embedding_uuid}": {
      "get": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Read embedding from collection",
        "description": "Returns a single embedding from a collection, looked up by its UUID.",
        "operationId": "get_embedding_from_collection_vectorstore_collections__collection_name__embeddings__embedding_uuid__get",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "embedding_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Embedding Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleEmbeddingResult"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "404": {
            "description": "Collection or embedding not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Update embedding in collection",
        "description": "Replaces the vector and/or metadata of an existing embedding, looked up by its UUID. A new vector must match the collection's dimensions.",
        "operationId": "update_embedding_in_collection_vectorstore_collections__collection_name__embeddings__embedding_uuid__put",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "embedding_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Embedding Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmbeddingBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleEmbeddingResult"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection or embedding not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Delete embedding from collection",
        "description": "Permanently deletes a single embedding from a collection, looked up by its UUID. This cannot be undone.",
        "operationId": "delete_embedding_vectorstore_collections__collection_name__embeddings__embedding_uuid__delete",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "embedding_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Embedding Uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation, no content returned"
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "404": {
            "description": "Collection or embedding not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vectorstore/collections/{collection_name}/embeddings": {
      "put": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Create or update embeddings in collection",
        "description": "Creates one or more embeddings in a collection, replacing any that already exist with the same UUID. Every vector must match the collection's dimensions.\n\nAuthorization is resolved per request:\n\n- If `authz` is set on the request body, that path is applied to all embeddings in the request.\n- Otherwise the collection's own path, `/vectorstore/collections/{collection_name}`, is used.\n- You need both `create` and `update` permission on the collection path and on the resulting embedding paths.",
        "operationId": "put_embeddings_in_collection_vectorstore_collections__collection_name__embeddings_put",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "ai_model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Model"
            }
          },
          {
            "name": "exclude_info",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Exclude Info"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmbeddingsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Read all embeddings from collection",
        "description": "Returns the embeddings in a collection, one page at a time. Set `exclude_info=true` to omit the `info` block from each result.",
        "operationId": "list_embeddings_in_collection_vectorstore_collections__collection_name__embeddings_get",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "exclude_info",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Exclude Info"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 100,
              "default": 100,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEmbeddingResponse"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Create embeddings in collection",
        "description": "Creates one or more embeddings in a collection. Every vector must match the collection's dimensions. Use `PUT` on this path instead if you want existing embeddings to be replaced rather than rejected.\n\nAuthorization is resolved per request:\n\n- If `authz` is set on the request body, those paths are applied to all embeddings in the request.\n- Otherwise the collection's own path, `/vectorstore/collections/{collection_name}`, is used.\n- You need `create` permission on the collection path and on the resulting embedding paths.",
        "operationId": "create_embeddings_in_collection_vectorstore_collections__collection_name__embeddings_post",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "ai_model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai Model"
            }
          },
          {
            "name": "exclude_info",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Exclude Info"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmbeddingsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/vectorstore/collections/{collection_name}": {
      "get": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Read collection info",
        "description": "Returns the metadata for a single collection, including its dimensions and vector type. Set `counts=true` to also get the number of embeddings available in it.",
        "operationId": "get_collection_vectorstore_collections__collection_name__get",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "counts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Counts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModel"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Update collection info",
        "description": "Updates the mutable metadata on a collection, such as its description. The collection's name, dimensions, and vector type cannot be changed after creation.",
        "operationId": "update_collection_vectorstore_collections__collection_name__patch",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Delete collection",
        "description": "Permanently deletes a collection and every embedding stored in it. This cannot be undone.",
        "operationId": "delete_collection_vectorstore_collections__collection_name__delete",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation, no content returned"
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEmbeddingsBody": {
        "properties": {
          "authz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authz"
          },
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingToCreate"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "embeddings"
        ],
        "title": "CreateEmbeddingsBody",
        "description": "Data for creating embeddings in a collection.\nauthz example: \"authz\": \"/vectorstore/collections/my_collection\""
      },
...
      "EmbeddingResponse": {
        "properties": {
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/SingleEmbeddingResult"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "embeddings"
        ],
        "title": "EmbeddingResponse",
        "description": "Embeddings from a single, already-known collection."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedEmbeddingResponse": {
        "properties": {
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/SingleEmbeddingResult"
            },
            "type": "array",
            "title": "Embeddings"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "next_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page"
          },
          "prev_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Page"
          }
        },
        "type": "object",
        "required": [
          "embeddings",
          "page",
          "page_size"
        ],
        "title": "PaginatedEmbeddingResponse",
        "description": "A page of embeddings from one collection."
      },
...
      "SingleEmbeddingResult": {
        "properties": {
          "vector": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Vector"
          },
          "input_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Index"
          },
          "embedding_id": {
            "type": "string",
            "format": "uuid",
            "title": "Embedding Id"
          },
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmbeddingInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "vector",
          "embedding_id"
        ],
        "title": "SingleEmbeddingResult",
        "description": "One embedding returned as a JSON array of floats."
      },
...
      "UpdateCollectionBody": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "UpdateCollectionBody",
        "description": "Request body for updating mutable properties of a Collection."
      },
      "UpdateEmbeddingBody": {
        "properties": {
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "authz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authz"
          }
        },
        "type": "object",
        "title": "UpdateEmbeddingBody",
        "description": "Request body for updating an embedding."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}